move provider display names to separate constants file
